varias alteracoes

+++ src/main/java/br/com/promocaodiaria/integrador/dto/ProdutoPromoDiariaDto.java
@@ -1,6 +1,5 @@
 package br.com.promocaodiaria.integrador.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -11,11 +10,11 @@ public class ProdutoPromoDiariaDto {
 	private String idIdentificador;
 	private String nome;
 	private String descricao;
-	private BigDecimal valor;
-	private BigDecimal vlPromocao;
+	private double valor;
+	private double vlPromocao;
 	private String codBarra;
 	private String codNcm;
-	private BigDecimal qtAtual;
+	private double qtAtual;
 	private String uniMedida;
 	private LocalDate dtFim;
 	private LocalDate dtInicio;
